# generated/swagger.json
{
	"components": {
...
		"schemas": {
...
			"ApiResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"error": {
						"type": "string"
					},
					"data": {}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Exercise": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"category": {
						"type": "string"
					},
					"difficulty": {
						"type": "string"
					},
					"equipment": {
						"type": "string"
					},
					"instructions": {
						"type": "string"
					},
					"link": {
						"type": "string"
					},
					"muscles_targeted": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"category",
					"difficulty",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ExercisesResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"error": {
						"type": "string"
					},
					"data": {},
					"exercises": {
						"items": {
							"$ref": "#/components/schemas/Exercise"
						},
						"type": "array"
					}
				},
				"required": [
					"success",
					"exercises"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ExerciseResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"error": {
						"type": "string"
					},
					"data": {},
					"exercise": {
						"$ref": "#/components/schemas/Exercise"
					}
				},
				"required": [
					"success",
					"exercise"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ExerciseLog": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"planDayExerciseId": {
						"type": "number",
						"format": "double"
					},
					"setsCompleted": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"repsCompleted": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"weightUsed": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"durationCompleted": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"roundsCompleted": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"restTimeTaken": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"timeTaken": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"isComplete": {
						"type": "boolean"
					},
					"isSkipped": {
						"type": "boolean"
					},
					"notes": {
						"type": "string",
						"nullable": true
					},
					"difficulty": {
						"type": "string",
						"nullable": true
					},
					"rating": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"planDayExerciseId",
					"setsCompleted",
					"repsCompleted",
					"weightUsed",
					"durationCompleted",
					"roundsCompleted",
					"restTimeTaken",
					"timeTaken",
					"isComplete",
					"isSkipped",
					"notes",
					"difficulty",
					"rating",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ExerciseLogResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"error": {
						"type": "string"
					},
					"data": {},
					"log": {
						"$ref": "#/components/schemas/ExerciseLog"
					}
				},
				"required": [
					"success",
					"log"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"WorkoutLog": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"workoutId": {
						"type": "number",
						"format": "double"
					},
					"totalTimeMinutes": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"daysCompleted": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"totalDays": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"totalVolume": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"averageRating": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"completedExercises": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array",
						"nullable": true
					},
					"completedBlocks": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array",
						"nullable": true
					},
					"completedDays": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array",
						"nullable": true
					},
					"isComplete": {
						"type": "boolean"
					},
					"isActive": {
						"type": "boolean"
					},
					"notes": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"workoutId",
					"totalTimeMinutes",
					"daysCompleted",
					"totalDays",
					"totalVolume",
					"averageRating",
					"completedExercises",
					"completedBlocks",
					"completedDays",
					"isComplete",
					"isActive",
					"notes",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"WorkoutLogResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"error": {
						"type": "string"
					},
					"data": {},
					"log": {
						"$ref": "#/components/schemas/WorkoutLog"
					}
				},
				"required": [
					"success",
					"log"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ExerciseDetails": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"muscleGroups": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"equipment": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"difficulty": {
						"type": "string"
					},
					"instructions": {
						"type": "string"
					},
					"link": {
						"type": "string"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"name",
					"description",
					"muscleGroups",
					"equipment",
					"difficulty",
					"instructions",
					"created_at",
					"updated_at"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PersonalRecord": {
				"properties": {
					"maxWeight": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"maxReps": {
						"type": "number",
						"format": "double"
					},
					"maxSets": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"maxWeight",
					"maxReps",
					"maxSets"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ExerciseUserStats": {
				"properties": {
					"totalAssignments": {
						"type": "number",
						"format": "double"
					},
					"totalCompletions": {
						"type": "number",
						"format": "double"
					},
					"completionRate": {
						"type": "number",
						"format": "double"
					},
					"averageSets": {
						"type": "number",
						"format": "double"
					},
					"averageReps": {
						"type": "number",
						"format": "double"
					},
					"averageWeight": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"lastPerformed": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"personalRecord": {
						"$ref": "#/components/schemas/PersonalRecord"
					}
				},
				"required": [
					"totalAssignments",
					"totalCompletions",
					"completionRate",
					"averageSets",
					"averageReps",
					"averageWeight",
					"lastPerformed",
					"personalRecord"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ExerciseSearchResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"error": {
						"type": "string"
					},
					"data": {},
					"exercise": {
						"$ref": "#/components/schemas/ExerciseDetails"
					},
					"userStats": {
						"allOf": [
							{
								"$ref": "#/components/schemas/ExerciseUserStats"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"success",
					"exercise",
					"userStats"
				],
				"type": "object",
				"additionalProperties": false
			},
			"WorkoutBlockWithExercise": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"workoutBlockId": {
						"type": "number",
						"format": "double"
					},
					"exerciseId": {
						"type": "number",
						"format": "double"
					},
					"sets": {
						"type": "number",
						"format": "double"
					},
					"reps": {
						"type": "number",
						"format": "double"
					},
					"weight": {
						"type": "number",
						"format": "double"
					},
					"duration": {
						"type": "number",
						"format": "double"
					},
					"restTime": {
						"type": "number",
						"format": "double"
					},
					"completed": {
						"type": "boolean"
					},
					"notes": {
						"type": "string"
					},
					"order": {
						"type": "number",
						"format": "double"
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					},
					"updated_at": {
						"type": "string",
						"format": "date-time"
					},
					"exercise": {
						"$ref": "#/components/schemas/Exercise"
					}
				},
				"required": [
					"id",
					"workoutBlockId",
					"exerciseId",
					"completed",
					"created_at",
					"updated_at",
					"exercise"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"WorkoutBlockExerciseResponse": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"error": {
						"type": "string"
					},
					"data": {},
					"workoutBlockExercise": {
						"$ref": "#/components/schemas/WorkoutBlockWithExercise"
					}
				},
				"required": [
					"success",
					"workoutBlockExercise"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreatePlanDayExerciseRequest": {
				"properties": {
					"workoutBlockId": {
						"type": "number",
						"format": "double"
					},
					"planDayId": {
						"type": "number",
						"format": "double"
					},
					"exerciseId": {
						"type": "number",
						"format": "double"
					},
					"sets": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"reps": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"weight": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"duration": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"restTime": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"completed": {
						"type": "boolean",
						"nullable": true
					},
					"notes": {
						"type": "string"
					},
					"order": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"workoutBlockId",
					"exerciseId",
					"sets",
					"reps",
					"weight",
					"duration",
					"restTime",
					"completed"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_InsertPlanDayExercise_": {
				"properties": {
					"weight": {},
					"order": {},
					"createdAt": {},
					"updatedAt": {},
					"completed": {},
					"workoutBlockId": {},
					"exerciseId": {},
					"sets": {},
					"reps": {},
					"duration": {},
					"restTime": {},
					"notes": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			}
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "mastersfit-server",
		"version": "1.0.0",
		"description": "RESTful API server for Masters Fit fitness tracking application",
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/exercises": {
			"get": {
				"operationId": "GetExercises",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ExercisesResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ExercisesResponse"
								}
							}
						}
					}
				},
				"description": "Get all exercises",
				"tags": [
					"Exercises"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": []
			},
			"post": {
				"operationId": "CreateExercise",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ExerciseResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Create new exercise",
				"tags": [
					"Exercises"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Exercise data",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"description": "Exercise data"
							}
						}
					}
				}
			}
		},
		"/exercises/{id}": {
			"get": {
				"operationId": "GetExercise",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ExerciseResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ExerciseResponse"
								}
							}
						}
					}
				},
				"description": "Get exercise by ID",
				"tags": [
					"Exercises"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "Exercise ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateExercise",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ExerciseResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ExerciseResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Update exercise",
				"tags": [
					"Exercises"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "Exercise ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "Updated exercise data",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"description": "Updated exercise data"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteExercise",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Delete exercise",
				"tags": [
					"Exercises"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "Exercise ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/exercises/{id}/link": {
			"put": {
				"operationId": "UpdateExerciseLink",
				"responses": {
					"200": {
						"description": "Link updated successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					},
					"404": {
						"description": "Exercise not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Update exercise link (YouTube video or image URL)",
				"tags": [
					"Exercises"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "Exercise ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "Link update data",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"link": {
										"type": "string",
										"nullable": true
									}
								},
								"required": [
									"link"
								],
								"type": "object",
								"description": "Link update data"
							}
						}
					}
				}
			}
		},
		"/logs/exercise": {
			"post": {
				"operationId": "CreateExerciseLog",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ExerciseLogResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Create exercise log",
				"tags": [
					"Logs"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "Exercise log data",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"description": "Exercise log data"
							}
						}
					}
				}
			}
		},
		"/logs/exercise/{exerciseLogId}": {
			"put": {
				"operationId": "UpdateExerciseLog",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ExerciseLogResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Update exercise log",
				"tags": [
					"Logs"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "Exercise log ID",
						"in": "path",
						"name": "exerciseLogId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "Update data",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"description": "Update data"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetExerciseLog",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ExerciseLogResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Get exercise log by ID",
				"tags": [
					"Logs"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "Exercise log ID",
						"in": "path",
						"name": "exerciseLogId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/logs/workout/{workoutId}/exercise/{planDayExerciseId}": {
			"post": {
				"operationId": "AddCompletedExercise",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WorkoutLogResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Add completed exercise to workout",
				"tags": [
					"Logs"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "Workout ID",
						"in": "path",
						"name": "workoutId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Plan day exercise ID",
						"in": "path",
						"name": "planDayExerciseId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/search/exercise/{userId}/{exerciseId}": {
			"get": {
				"operationId": "SearchExercise",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ExerciseSearchResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					},
					"404": {
						"description": "Not Found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Search for exercise details and user stats",
				"tags": [
					"Search"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "User ID",
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "Exercise ID",
						"in": "path",
						"name": "exerciseId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/search/exercises": {
			"get": {
				"operationId": "SearchExercises",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"exercises": {
											"items": {},
											"type": "array"
										},
										"success": {
											"type": "boolean"
										}
									},
									"required": [
										"exercises",
										"success"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse"
								}
							}
						}
					}
				},
				"description": "Search exercises by name or muscle group",
				"tags": [
					"Search"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "Search query",
						"in": "query",
						"name": "query",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/workouts/plan-day/{planDayId}/exercise": {
			"post": {
				"operationId": "CreatePlanDayExercise",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WorkoutBlockExerciseResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WorkoutBlockExerciseResponse"
								}
							}
						}
					}
				},
				"description": "Create plan day exercise",
				"tags": [
					"Workouts"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "Plan day ID",
						"in": "path",
						"name": "planDayId",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "Exercise data",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreatePlanDayExerciseRequest",
								"description": "Exercise data"
							}
						}
					}
				}
			}
		},
		"/workouts/exercise/{id}": {
			"put": {
				"operationId": "UpdatePlanDayExercise",
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WorkoutBlockExerciseResponse"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/WorkoutBlockExerciseResponse"
								}
							}
						}
					}
				},
				"description": "Update plan day exercise",
				"tags": [
					"Workouts"
				],
				"security": [
					{
						"bearerAuth": []
					}
				],
				"parameters": [
					{
						"description": "Exercise ID",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"description": "Updated exercise data",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_InsertPlanDayExercise_",
								"description": "Updated exercise data"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}